New files!
The first version of this library. Encouraged by my friend, I decided to upload the main file "PythonFileRewriter" with a small custom made HTTP server. This file is the answer to the connection between files and python. Of course, this option already exists in some framework libraries as Flask Of CherryPy but I decided to focus only on this and on user experience.